requirements.txt
#zip
webdriver-manager
streamlit
Pillow
google-cloud-aiplatform>=1.29.0
selenium-screenshot
seleniumbase
xlsxwriter
pandas
selenium
google-generativeai
pyvirtualdisplay
google-cloud-aiplatform 
google-cloud-vision
google-cloud-storage 
google-cloud-translate 
google-cloud-automl
vertexai
#seleniumbase
webdriver-manager
#chromedriver-binary